# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MCQ Practice App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a specific course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}/subjects:
    get:
      operationId: getSubjectsByCourse
      tags: [subjects]
      summary: Get subjects for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
...
  /admin/courses:
    get:
      operationId: getAdminCourses
      tags: [admin]
      summary: Get all courses (admin view with full details)
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [admin]
      summary: Create a new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /admin/courses/{courseId}:
    put:
      operationId: updateCourse
      tags: [admin]
      summary: Update a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [admin]
      summary: Delete a course and all its content
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Course:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        icon:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        isActive:
          type: boolean
        orderNumber:
          type: integer
        subjectCount:
          type: ["integer", "null"]
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - code
        - createdAt
    CourseInput:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
        description:
          type: string
        icon:
          type: string
        color:
          type: string
      required:
        - name
        - code
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        levelId:
          type: ["integer", "null"]
        name:
          type: string
        code:
          type: string
        description:
          type: ["string", "null"]
        isActive:
          type: boolean
        chapterCount:
          type: ["integer", "null"]
        questionCount:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - courseId
        - name
        - code
        - createdAt
...